backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.3.0
python-multipart==0.0.6
pydantic==2.4.2
